# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AskIn": {
        "additionalProperties": false,
        "properties": {
          "client_message_id": {
            "description": "Created by the client; a repeat is ignored.",
            "format": "uuid",
            "title": "Client Message Id",
            "type": "string"
          },
          "comparison": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ComparisonIn"
              },
              {
                "type": "null"
              }
            ]
          },
          "content": {
            "maxLength": 16000,
            "title": "Content",
            "type": "string"
          },
          "effort": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Effort"
              },
              {
                "type": "null"
              }
            ]
          },
          "locale": {
            "$ref": "#/components/schemas/Locale"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "style": {
            "$ref": "#/components/schemas/AnswerStyle",
            "default": "concise"
          }
        },
        "required": [
          "client_message_id",
          "content",
          "model",
          "locale"
        ],
        "title": "AskIn",
        "type": "object"
      },
...
      "ChatEnvelopeOut": {
        "properties": {
          "chat": {
            "$ref": "#/components/schemas/ChatOut"
          }
        },
        "required": [
          "chat"
        ],
        "title": "ChatEnvelopeOut",
        "type": "object"
      },
...
      "ChatListOut": {
        "properties": {
          "chats": {
            "items": {
              "$ref": "#/components/schemas/ChatListItemOut"
            },
            "title": "Chats",
            "type": "array"
          }
        },
        "required": [
          "chats"
        ],
        "title": "ChatListOut",
        "type": "object"
      },
...
      "ChatScope": {
        "description": "`all`: every `ready` document at question time. `selected`: the chosen ones.",
        "enum": [
          "all",
          "selected"
        ],
        "title": "ChatScope",
        "type": "string"
      },
...
      "CreateChatIn": {
        "additionalProperties": false,
        "properties": {
          "document_ids": {
            "description": "Only for `selected`.",
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Document Ids",
            "type": "array"
          },
          "scope": {
            "$ref": "#/components/schemas/ChatScope",
            "default": "all"
          }
        },
        "title": "CreateChatIn",
        "type": "object"
      },
...
      "DocumentListOut": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentOut"
            },
            "title": "Documents",
            "type": "array"
          }
        },
        "required": [
          "documents"
        ],
        "title": "DocumentListOut",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "MessageListOut": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageOut"
            },
            "title": "Messages",
            "type": "array"
          }
        },
        "required": [
          "messages"
        ],
        "title": "MessageListOut",
        "type": "object"
      },
...
      "RegenerateIn": {
        "additionalProperties": false,
        "properties": {
          "effort": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Effort"
              },
              {
                "type": "null"
              }
            ]
          },
          "locale": {
            "$ref": "#/components/schemas/Locale"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Default: the answer's model.",
            "title": "Model"
          },
          "style": {
            "$ref": "#/components/schemas/AnswerStyle",
            "default": "concise"
          }
        },
        "required": [
          "locale"
        ],
        "title": "RegenerateIn",
        "type": "object"
      },
...
      "StopIn": {
        "additionalProperties": false,
        "properties": {
          "lane": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Lane"
              },
              {
                "type": "null"
              }
            ],
            "description": "Without a lane every lane stops."
          }
        },
        "title": "StopIn",
        "type": "object"
      },
      "StopOut": {
        "properties": {
          "stopped": {
            "items": {
              "$ref": "#/components/schemas/Lane"
            },
            "title": "Stopped",
            "type": "array"
          }
        },
        "required": [
          "stopped"
        ],
        "title": "StopOut",
        "type": "object"
      },
...
      "UpdateChatIn": {
        "additionalProperties": false,
        "properties": {
          "document_ids": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Ids"
          },
          "scope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatScope"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "1 to 120 characters.",
            "title": "Title"
          }
        },
        "title": "UpdateChatIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Siteco Document Chat API",
    "version": "contract"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/chats": {
      "get": {
        "description": "All chats, most recently active first.",
        "operationId": "list_chats_api_chats_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatListOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Chats",
        "tags": [
          "chats"
        ]
      },
      "post": {
        "operationId": "create_chat_api_chats_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChatIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatEnvelopeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Create Chat",
        "tags": [
          "chats"
        ]
      }
    },
    "/api/chats/{chat_id}": {
      "delete": {
        "description": "Stops running answers of the chat first.",
        "operationId": "delete_chat_api_chats__chat_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chat Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Delete Chat",
        "tags": [
          "chats"
        ]
      },
      "get": {
        "operationId": "get_chat_api_chats__chat_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chat Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatEnvelopeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Chat",
        "tags": [
          "chats"
        ]
      },
      "patch": {
        "description": "A new title is the user's and is never replaced automatically.",
        "operationId": "update_chat_api_chats__chat_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chat Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChatIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatEnvelopeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Update Chat",
        "tags": [
          "chats"
        ]
      }
    },
    "/api/chats/{chat_id}/attachments": {
      "get": {
        "description": "Documents uploaded into this chat, newest first. Poll while any is processing.",
        "operationId": "list_attachments_api_chats__chat_id__attachments_get",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chat Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "List Attachments",
        "tags": [
          "chats"
        ]
      }
    },
    "/api/chats/{chat_id}/attachments/{document_id}": {
      "delete": {
        "description": "Removes the document from this chat. One that is neither in the library nor in another\nchat is deleted completely (file, index, cited text).",
        "operationId": "remove_attachment_api_chats__chat_id__attachments__document_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chat Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Remove Attachment",
        "tags": [
          "chats"
        ]
      }
    },
    "/api/chats/{chat_id}/messages": {
      "get": {
        "description": "All messages, oldest first.",
        "operationId": "list_messages_api_chats__chat_id__messages_get",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chat Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "List Messages",
        "tags": [
          "chats"
        ]
      },
      "post": {
        "description": "Asks a question and streams the answer. Without an API key the stream carries the\nsources only (`done.status = sources_only`).",
        "operationId": "ask_api_chats__chat_id__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chat Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "itemSchema": {
                  "properties": {
                    "data": {
                      "type": "string"
                    },
                    "event": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "retry": {
                      "minimum": 0,
                      "type": "integer"
                    }
                  },
                  "type": "object"
                },
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Events `meta`, `status`, `sources`, `delta`, `citation`, then exactly one `done` or `error` (payloads: Sse* schemas). A `: ping` comment every 15 s."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Refused before the stream opened."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Refused before the stream opened."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Refused before the stream opened."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Refused before the stream opened."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Refused before the stream opened."
          }
        },
        "summary": "Ask",
        "tags": [
          "chats"
        ]
      }
    },
    "/api/chats/{chat_id}/messages/{assistant_id}/prefer": {
      "post": {
        "description": "Keeps this answer of a comparison; only the kept answer goes into later history.",
        "operationId": "prefer_api_chats__chat_id__messages__assistant_id__prefer_post",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chat Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "assistant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Assistant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Prefer",
        "tags": [
          "chats"
        ]
      }
    },
    "/api/chats/{chat_id}/messages/{assistant_id}/regenerate": {
      "post": {
        "description": "Replaces an answer to the latest question (same lane, same message id).",
        "operationId": "regenerate_api_chats__chat_id__messages__assistant_id__regenerate_post",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chat Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "assistant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Assistant Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegenerateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "itemSchema": {
                  "properties": {
                    "data": {
                      "type": "string"
                    },
                    "event": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "retry": {
                      "minimum": 0,
                      "type": "integer"
                    }
                  },
                  "type": "object"
                },
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Events `meta`, `status`, `sources`, `delta`, `citation`, then exactly one `done` or `error` (payloads: Sse* schemas). A `: ping` comment every 15 s."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Refused before the stream opened."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Refused before the stream opened."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Refused before the stream opened."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Refused before the stream opened."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Refused before the stream opened."
          }
        },
        "summary": "Regenerate",
        "tags": [
          "chats"
        ]
      }
    },
    "/api/chats/{chat_id}/stop": {
      "post": {
        "description": "Second safety net next to aborting the request: stops the lane's answer, which is then\nsaved as `stopped`.",
        "operationId": "stop_api_chats__chat_id__stop_post",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chat Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/StopIn"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StopOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Stop",
        "tags": [
          "chats"
        ]
      }
    },
...
  }
}